+++ NEWS.md
@@ -1,6 +1,7 @@
 # idaifieldR 0.3.4 _2023-11-24_
 * Make `get_field_index()` more efficient by utilizing vectors instead of `lapply()`. 
 * Adding *liesWithinLayer* in `get_field_index()` and `simplify_idaifield()` is now an option. This way, in `get_field_index()` the "shortDescription" can be added without the more time intensive recursive layer search that comes with the now available option `find_layers = TRUE`. For info on what this does, see documentation of `find_layer()`.
+* `find_layer()` can now handle a vector of multiple ids.
 
 # idaifieldR 0.3.3 _2023-11-23_
 ### New features